SF: Add retry to EventThread postEvent
When consumer BitTube is under heavy Event load, like constantly
updating use case providing VSYNC events, it is possible for Hotplug
events to fail sending across the socket. In such case, retry
mechanism is needed to send it again. We would like not to invoke
retry logic so we are also doubling the BitTube size to avoid
Event backlog in the first place.

Change-Id: I7bcad9512cb6ecdac945fd765eb9258ccd88fee0
CRs-Fixed: 2700045